pyproject.toml
[project]
name = "packastack"
version = "0.1.0"
description = "A tool for building OpenStack packages for Ubuntu"
readme = "README.md"
license = "GPL-3.0-only"
requires-python = ">=3.11"
authors = [
    { name = "Canonical Ltd." }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "python-debian>=0.1.49",
]

[project.scripts]
packastack = "packastack.cli:app"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0",
    "responses>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/packastack"]

[tool.ruff]
src = ["src", "tests"]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "ARG",    # Unused arguments are common in pytest fixtures
    "SIM117", # Nested with statements are often clearer in tests
]

[tool.ruff.lint.isort]
known-first-party = ["packastack"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src/packastack", "tests"]

[[tool.mypy.overrides]]
module = ["git", "git.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["debian", "debian.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-v",
    "--cov=packastack",
    "--cov-report=term-missing",
    "--cov-fail-under=95",
]

[tool.coverage.run]
source = ["src/packastack"]
branch = true
omit = ["src/packastack/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "pragma: nocover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
